+++ apps/server/src/modules/news/uc/index.ts
@@ -1,2 +1,1 @@
 export * from './news.uc';
-export * from './create-news.dto';