+++ apps/server/src/modules/task/entity/task.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, ManyToOne, Property } from '@mikro-orm/core';
-import { BaseEntityWithTimestamps } from '../../../shared/domain';
+import { BaseEntityWithTimestamps } from '@shared/domain';
 import { Course } from './course.entity';
 import { Lesson } from './lesson.entity';
 
@@ -7,27 +7,16 @@ import { Lesson } from './lesson.entity';
 export class Task extends BaseEntityWithTimestamps {
 	@Property()
 	name: string;
+
 	@Property()
 	dueDate?: Date;
+
 	@Property()
 	private?: boolean;
 
 	@ManyToOne({ fieldName: 'courseId' })
 	course: Course;
+
 	@ManyToOne({ fieldName: 'lessonId' })
 	lesson?: Lesson;
-
-	constructor() {
-		super();
-	}
-}
-
-export class ITaskMetadata extends BaseEntityWithTimestamps {
-	name: string;
-
-	duedate: Date;
-
-	courseName: string;
-
-	displayColor: string;
 }