+++ apps/server/src/shared/domain/entity/base.entity.ts
@@ -1,7 +1,8 @@
-import { PrimaryKey, Property, SerializedPrimaryKey } from '@mikro-orm/core';
+import { Entity, PrimaryKey, Property, SerializedPrimaryKey } from '@mikro-orm/core';
 import { ObjectId } from '@mikro-orm/mongodb';
 
-export abstract class BaseEntity {
+@Entity()
+export class BaseEntity {
 	@PrimaryKey()
 	_id!: ObjectId;
 
@@ -9,7 +10,11 @@ export abstract class BaseEntity {
 	id!: string;
 }
 
-export abstract class BaseEntityWithTimestamps {
+// NOTE we have to include BaseEntityWithTimestamps in the entity discovery if we inherit from BaseEntity.
+// that can be cumbersome e.g. in tests. that's why we define it as a root class here.
+// TODO check if we can use EntitySchema to prevent code duplication (decorators don't work for defining properties btw.)
+@Entity()
+export class BaseEntityWithTimestamps extends BaseEntity {
 	@PrimaryKey()
 	_id!: ObjectId;